# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple version of user profile using Echo app.",
        "title": "Go Profile",
        "contact": {
            "name": "Vidya Haikal",
            "email": "haikalvidya@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/login": {
            "post": {
                "description": "Perform user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Actions"
                ],
                "summary": "Authenticate a user",
                "operationId": "user-login",
                "parameters": [
                    {
                        "description": "User's credentials",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/profiles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the list of all profiles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles Actions"
                ],
                "summary": "Get profiles",
                "operationId": "profiles-get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.ProfileResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles Actions"
                ],
                "summary": "Create profile",
                "operationId": "profiles-create",
                "parameters": [
                    {
                        "description": "Profile alamat, pekerjaan, nama lengkap, pendidikan, nomor telpon",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.Data"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/profiles/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get the profiles by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles Actions"
                ],
                "summary": "Get profile",
                "operationId": "profiles-get-one",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.ProfileResponse"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles Actions"
                ],
                "summary": "Update profile",
                "operationId": "profiles-update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Profile alamat, pekerjaan, nama lengkap, pendidikan, nomor telpon",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Data"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete profile",
                "tags": [
                    "Profiles Actions"
                ],
                "summary": "Delete profile",
                "operationId": "profiles-delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Profile ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/responses.Data"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/refresh": {
            "post": {
                "description": "Perform refresh access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Actions"
                ],
                "summary": "Refresh access token",
                "operationId": "user-refresh",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.LoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/register": {
            "post": {
                "description": "New user registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Actions"
                ],
                "summary": "Register",
                "operationId": "user-register",
                "parameters": [
                    {
                        "description": "User's email, user's password",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.Data"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "requests.CreateProfileRequest": {
            "type": "object",
            "required": [
                "alamat",
                "nama_lengkap",
                "nomer_telpon",
                "pekerjaan",
                "pendidikan_terakhir"
            ],
            "properties": {
                "alamat": {
                    "type": "string",
                    "example": "Jl. Echo"
                },
                "nama_lengkap": {
                    "type": "string",
                    "example": "Echo Parecho"
                },
                "nomer_telpon": {
                    "type": "string",
                    "example": "+6281234123421"
                },
                "pekerjaan": {
                    "type": "string",
                    "example": "DevOps"
                },
                "pendidikan_terakhir": {
                    "type": "string",
                    "example": "Mahasiswa"
                }
            }
        },
        "requests.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123456"
                }
            }
        },
        "requests.RefreshRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "example": "refresh_token"
                }
            }
        },
        "requests.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "JohnDoe"
                },
                "password": {
                    "type": "string",
                    "example": "password123456"
                }
            }
        },
        "requests.UpdateProfileRequest": {
            "type": "object",
            "required": [
                "alamat",
                "nama_lengkap",
                "nomer_telpon",
                "pekerjaan",
                "pendidikan_terakhir"
            ],
            "properties": {
                "alamat": {
                    "type": "string",
                    "example": "Jl. Echo"
                },
                "nama_lengkap": {
                    "type": "string",
                    "example": "Echo Parecho"
                },
                "nomer_telpon": {
                    "type": "string",
                    "example": "+6281234123421"
                },
                "pekerjaan": {
                    "type": "string",
                    "example": "DevOps"
                },
                "pendidikan_terakhir": {
                    "type": "string",
                    "example": "Mahasiswa"
                }
            }
        },
        "responses.Data": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "responses.LoginResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "exp": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "responses.ProfileResponse": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string",
                    "example": "Jl. Echo"
                },
                "email": {
                    "type": "string",
                    "example": "John@doo.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nama_lengkap": {
                    "type": "string",
                    "example": "Echo Parecho"
                },
                "nomer_telpon": {
                    "type": "string",
                    "example": "+6281234123421"
                },
                "pekerjaan": {
                    "type": "string",
                    "example": "DevOps"
                },
                "pendidikan_terakhir": {
                    "type": "string",
                    "example": "Mahasiswa"
                },
                "username": {
                    "type": "string",
                    "example": "John"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}